+++ community/src/main/java/org/dhis2/community/tasking/ui/TaskingUi.kt
@@ -230,121 +230,4 @@ fun TaskingUi(
     }
 }
 
-@Composable
-fun TaskItem(task: TaskingUiModel, onTaskClick: (TaskingUiModel) -> Unit)  {
-    val statusColor = task.status.color
-    val infoList = listOf(
-        AdditionalInfoItem(
-            icon = {
-                Icon(
-                    imageVector = Icons.Outlined.Flag,
-                    contentDescription = null,
-                    tint = task.priority.color,
-                    modifier = Modifier.size(24.dp)
-                )
-            },
-            value = task.priority.label,
-            color = task.priority.color,
-            truncate = false,
-            isConstantItem = false
-        ),
-        AdditionalInfoItem(
-            icon = {
-                Icon(
-                    imageVector = Icons.Outlined.Event,
-                    contentDescription = null,
-                    tint = statusColor,
-                    modifier = Modifier.size(24.dp)
-                )
-            },
-            value = task.status.label,
-            color = statusColor,
-            truncate = false,
-            isConstantItem = false
-        ),
-        AdditionalInfoItem(
-            icon = {
-                Icon(
-                    imageVector = Icons.Outlined.CheckCircle,
-                    contentDescription = null,
-                    tint = task.progressColor,
-                    modifier = Modifier.size(24.dp)
-                )
-            },
-            value = task.progressDisplay,
-            color = task.progressColor,
-            truncate = false,
-            isConstantItem = false
-        ),
-        AdditionalInfoItem(
-            key = "Name",
-            value = task.teiPrimary,
-            color = TextColor.OnSurface,
-            truncate = false,
-            isConstantItem = false
-        ),
-        AdditionalInfoItem(
-            key = "Village Name",
-            value = task.teiSecondary,
-            color = TextColor.OnSurface,
-            truncate = false,
-            isConstantItem = false
-        ),
-        AdditionalInfoItem(
-            key = "Date of Birth",
-            value = task.teiTertiary,
-            color = TextColor.OnSurface,
-            truncate = false,
-            isConstantItem = false
-        )
-    )
-    val additionalInfoState = rememberAdditionalInfoColumnState(
-        additionalInfoList = infoList,
-        syncProgressItem = AdditionalInfoItem(
-            value = "",
-            color = TextColor.OnSurface,
-            isConstantItem = true
-        ),
-        expandLabelText = "Show more",
-        shrinkLabelText = "Show less",
-        minItemsToShow = 3,
-        scrollableContent = false
-    )
-    val listCardState = rememberListCardState(
-        title = ListCardTitleModel(
-            text = task.taskName,
-            color = SurfaceColor.Primary
-        ),
-        description = ListCardDescriptionModel(
-            text = task.displayProgramName,
-            color = TextColor.OnSurface
-        ),
-        lastUpdated = task.dueDate?.let {
-            "Due: ${SimpleDateFormat("dd/MM/yyyy", Locale.getDefault()).format(it)}"
-        },
-        additionalInfoColumnState = additionalInfoState,
-        loading = false,
-        shadow = true,
-        itemVerticalPadding = 8.dp,
-        selectionState = SelectionState.NONE
-    )
-    ListCard(
-        modifier = Modifier.padding(8.dp),
-        listCardState = listCardState,
-        listAvatar = {
-            Avatar(
-                style = AvatarStyleData.Metadata(
-                    imageCardData = task.metadataIconData.imageCardData,
-                    avatarSize = MetadataAvatarSize.S(),
-                    tintColor = task.metadataIconData.color
-                )
-            )
-        },
-        onCardClick = {
-            onTaskClick(task)
-        },
-        onSizeChanged = {}
-    )
-}
-
 private enum class FilterSheetType { PROGRAM, PRIORITY, STATUS, DUE_DATE }
